Set up permissions for Books and applied checks in routes

+++ routes/books.js
@@ -3,6 +3,7 @@ const bodyParser = require('koa-bodyparser');
 const model = require('../models/books');
 const auth = require('../controllers/auth');
 const {validateBook, validateBookUpdate} = require('../controllers/validation');
+const can = require('../permissions/books');
 
 const router = Router({prefix: '/api/v1/books'});
 
@@ -38,33 +39,66 @@ async function getById(ctx) {
 
 async function createBook(ctx) {
     let body = ctx.request.body;
-    let result = await model.add(body);
-    if (result.affectedRows) {
-        ctx.body = {ID: result.insertId, created: true}
-        ctx.status = 201;
+    let permission = can.create(ctx.state.user);
+
+    if (!permission.granted) {
+        ctx.status = 403;
     } else {
-        ctx.status = 404;
+        let result = await model.add(body);
+        if (result.affectedRows) {
+            ctx.body = {ID: result.insertId, created: true}
+            ctx.status = 201;
+        } else {
+            ctx.status = 404;
+        }
     }
 }
 
 async function updateBook(ctx){
     let id = ctx.params.id;
-    let body = ctx.request.body;
-    let result = await model.updateById(body, id);
-    if (result.affectedRows) {
-        ctx.body = {ID: id, updated: true};
-        ctx.status = 200;
+    let result = await model.getById(id);
+
+    if (result.length) {
+        let data = result[0];
+        let permission = can.update(ctx.state.user);
+
+        if (!permission.granted) {
+            ctx.status = 403;
+        } else {
+            let data = ctx.request.body;
+            let result = await model.updateById(data, id);
+            if (result.affectedRows) {
+                ctx.body = {ID: id, updated: true};
+                ctx.status = 200;
+            } else {
+                ctx.status = 404;
+            }
+        }
     } else {
         ctx.status = 404;
     }
 }
 
-  async function deleteBook(ctx){
+async function deleteBook(ctx){
     let id = ctx.params.id;
-    let result = await model.deleteById(id);
-    if (result.affectedRows) {
-        ctx.body = {ID: id, deleted: true};
-        ctx.status = 200;
+    let result = await model.getById(id);
+
+    if (result.length) {
+        let data = result[0];
+        let permission = can.delete(ctx.state.user);
+
+        if (!permission.granted) {
+            ctx.status = 403;
+        } else {
+
+            let result = await model.deleteById(id);
+            if (result.affectedRows) {
+                ctx.body = {ID: id, deleted: true};
+                ctx.status = 200;
+            } else {
+                ctx.status = 404;
+            }
+        }       
     } else {
         ctx.status = 404;
     }
